redirect pages on the site
Redirect app supplied by django-redirect checks for any old paths that can be redirected to an appropriate location.

Early admin feature gives a link in the post editing form.

Test data was included.

Template for the change view form is inside `blog/templates`.

Settings includes django-redirects and APPEND_SLASH is turned on. This option is important for looking at URLs with no slash for redirection.