summernote plugin removed, post edit template added
Summernote removed from the content field of posts. The app reference has been totally stripped from tests too.

When posts are edited the change_form template is loaded to allow the end-users access to the url redirect creator - a field  located at the bottom that shows all the connected urls.