rework total_post_count, made param as optional, catches errors and added docstring
total_post_count takes a string and this is optional now as it defaults to DEFAULT_POST_STATUS defined as global variable inside the blog_tags.py

If you provide a wrong option, or give the a value type that's different to strings, you'll receive an error.

Docstring in method for the to improve our documentation.