Name change PostShoutout, was previously PostRedirect
The reason for postredirect to postshoutout name change was because postshoutout has a conflict of interest with the modules handling actuall HTTP redirections.

Develoeprs reading postredirect may correctly assume a 301 request is taking place, however this wasn't the case for PostRedirect models. So we decided to make the PostShoutout the offical name.

Post Shoutout are meant to convey the idea that a post is giving special mention to something or someone. For example, a Post Shoutout titled "My best friend's wedding gallery" is more obvious to lead to something specifically photo related.