contact page is no longer offered by blog_improved
This commit removes the contact url from the blog. Delivering a contact page was a previous feature of blog_improved.

Our belief directs us in thinking that web authors are already in control of their own pages, and when they install blog_improved then they expect blogging capabilities. We striving to keep the blog_improved to a narrow scope. For this reason, we are removing the contact url, however the template will remain and be maintained going forward.